+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## Changes in version 0.0.2 (in development)
 
+* Updated `tools/openapi.yaml` to rename some unintuitive names.
+  Regenerated and adjusted remaining code:
+  - renamed class `Response` into `ResponseType`
+  - renamed class `StatusCode` into `JobStatus`
+  - renamed class `Type` into `JobType`
+  - renamed class `Type1` into `DataType`
 
 ## Changes in version 0.0.1